[ConnectionManager] Remove stale connections.
Before this patch, netmine did not purge closed connections from the
manager's connection list, and so a flurry of broken pipe errors would
occur each time a message was sent. Now, when a message is sent and an
error occurs, the connection is regarded as closed and is removed from
the list of connections.